Add files via upload
Task: Given two sequences 𝑎1, 𝑎2, . . . , 𝑎𝑛(𝑎𝑖 is the profit per click of the 𝑖-th ad) and 𝑏 is the average number of clicks per day of the 𝑖-th slot), we need to partition them into 𝑛 pairs (𝑎) such that the sum of their products is maximized.

Input Format. The first line contains an integer 𝑛, the second one contains a sequence of integers 𝑎1, 𝑎2, . . . , 𝑎𝑛, the third one contains a sequence of integers 𝑏.

Constraints: 1 ≤ 𝑛 ≤ 10^3;  −10^5  ≤𝑎𝑖,  𝑏𝑖≤ 10^5; for all 1 ≤ i ≤ n

Output Format. Output the maximum value of ∑︀( i=1 to n)𝑎𝑖 *𝑐𝑖 where 𝑐𝑖 is the permutation of 𝑏𝑖